Add workaround for jQuery.validate 1.11 (clientside_validation 7.x-1.x).
The iframe elements rendered by braintree’s hosted fields set a
type-attribute on each iframe. jQuery.validate works on [type="number"]
though which means it tries to validate the iframe. This fails because
an iframe-element is not a form input (ie. doesn’t have a form
property).

That’s why we remove the type-attributes from the iframes as a
workaround.